refactor(overlap): split shape model into _overlap.ts
Keeps both modules under the 200 LoC guideline: _overlap.ts holds the shape
model (RouteShape, shapeOf, shapesOverlap), operations/overlap.ts the public
API and tree traversal.
