+++ src/operations/overlap.ts
@@ -1,6 +1,6 @@
-import { createRouter } from "../context.ts";
-import type { MatchedRoute, MethodData, Node, RouterContext } from "../types.ts";
-import { addRoute } from "./add.ts";
+import { mayMatchAt, routeToShapes, shapeOf, shapesOverlap } from "../_overlap.ts";
+import type { Edge, RouteShape } from "../_overlap.ts";
+import type { MatchedRoute, Node, RouterContext } from "../types.ts";
 
 /**
  * Whether two route patterns can match a common concrete path (their match-sets
@@ -26,11 +26,11 @@ import { addRoute } from "./add.ts";
  * routesOverlap("/a/**", "/a"); // true (`**` matches zero segments)
  */
 export function routesOverlap(patternA: string, patternB: string): boolean {
-  const a = _routeToShapes(patternA);
-  const b = _routeToShapes(patternB);
+  const a = routeToShapes(patternA);
+  const b = routeToShapes(patternB);
   for (const x of a) {
     for (const y of b) {
-      if (_shapesOverlap(x, y)) return true;
+      if (shapesOverlap(x, y)) return true;
     }
   }
   return false;
@@ -56,72 +56,10 @@ export function findOverlappingRoutes<T>(
   method: string = "",
   pattern: string,
 ): MatchedRoute<T>[] {
-  const query = _routeToShapes(pattern);
+  const query = routeToShapes(pattern);
   const matches: MatchedRoute<T>[] = [];
   _collectOverlaps(ctx.root, method, query, [], new Set(), matches);
   return matches;
-}
-
-/**
- * A canonical (fully expanded) route shape: fixed single-segment matchers
- * (`string` literal | `RegExp` constraint | `undefined` = any) followed by a
- * variable-length tail. The tail (trailing `*`, `**`, `**:name`) matches any
- * segment values, so it only constrains the total number of segments:
- * trailing bare `*` -> `[0, 1]`, `**` -> `[0, Infinity]`,
- * `**:name` -> `[1, Infinity]`, no variable tail -> `[0, 0]`.
- */
-interface RouteShape {
-  fixed: (string | RegExp | undefined)[];
-  tailMin: number;
-  tailMax: number;
-}
-
-/**
- * A radix-tree edge on the path from the root to a node: a static key (already
- * decoded, so it may be a literal `*`/`**`), or a param (`0`) / wildcard (`1`)
- * branch. Carrying the node kind keeps escaped-literal static keys
- * distinguishable from dynamic segments.
- */
-type Edge = string | 0 | 1;
-
-/**
- * Expand a route pattern into its canonical shapes by inserting it into a
- * throwaway router with the real `addRoute` pipeline (group delimiters, escape
- * encoding, modifiers) and reading the resulting tree entries. Both query
- * patterns and registered routes are therefore classified by the exact same
- * code, so overlap stays consistent with route matching by construction.
- */
-function _routeToShapes(pattern: string): RouteShape[] {
-  const ctx = createRouter();
-  addRoute(ctx, "", pattern);
-  const shapes: RouteShape[] = [];
-  _collectShapes(ctx.root, [], shapes);
-  return shapes;
-}
-
-function _collectShapes(node: Node, edges: Edge[], shapes: RouteShape[]): void {
-  if (node.methods) {
-    for (const entry of node.methods[""] || []) {
-      shapes.push(_shapeOf(edges, entry));
-    }
-  }
-  if (node.static) {
-    for (const key in node.static) {
-      edges.push(key);
-      _collectShapes(node.static[key], edges, shapes);
-      edges.pop();
-    }
-  }
-  if (node.param) {
-    edges.push(0);
-    _collectShapes(node.param, edges, shapes);
-    edges.pop();
-  }
-  if (node.wildcard) {
-    edges.push(1);
-    _collectShapes(node.wildcard, edges, shapes);
-    edges.pop();
-  }
 }
 
 function _collectOverlaps<T>(
@@ -146,7 +84,7 @@ function _collectOverlaps<T>(
   if (node.static) {
     for (const key in node.static) {
       // Static keys are value-constrained: skip subtrees the query can't reach.
-      if (_mayMatchAt(query, edges.length, key)) {
+      if (mayMatchAt(query, edges.length, key)) {
         edges.push(key);
         _collectOverlaps(node.static[key], method, query, edges, seen, matches);
         edges.pop();
@@ -158,110 +96,12 @@ function _collectOverlaps<T>(
     if (data) {
       for (const entry of data) {
         if (seen.has(entry.data)) continue;
-        const shape = _shapeOf(edges, entry);
-        if (query.some((q) => _shapesOverlap(q, shape))) {
+        const shape = shapeOf(edges, entry);
+        if (query.some((q) => shapesOverlap(q, shape))) {
           seen.add(entry.data);
           matches.push({ data: entry.data });
         }
       }
     }
   }
-}
-
-// A route entry's shape never changes once inserted; cache across queries.
-const _shapeCache = new WeakMap<MethodData, RouteShape>();
-
-/**
- * Build the {@link RouteShape} of a registered route entry from its tree
- * position (kind-tagged edges) and its own `paramsMap` classification.
- */
-function _shapeOf(edges: Edge[], entry: MethodData): RouteShape {
-  let shape = _shapeCache.get(entry);
-  if (shape) return shape;
-  const fixed: RouteShape["fixed"] = [];
-  let tailMin = 0;
-  let tailMax = 0;
-  const pMap = entry.paramsMap;
-  for (let d = 0; d < edges.length; d++) {
-    const edge = edges[d];
-    if (typeof edge === "string") {
-      fixed.push(edge);
-    } else if (edge === 1) {
-      // Wildcard is always terminal and its paramsMap entry is always last
-      // (`**` is optional, `**:name` requires one segment).
-      tailMin = pMap![pMap!.length - 1][2] ? 0 : 1;
-      tailMax = Number.POSITIVE_INFINITY;
-    } else if (pMap) {
-      // Param: classified by this entry's paramsMap entry at this segment index.
-      const p = pMap.find((e) => e[0] === d)!;
-      if (p[1] instanceof RegExp) {
-        fixed.push(p[1]);
-      } else if (p[2] /* bare `*` */ && d === edges.length - 1) {
-        // A trailing bare `*` matches zero-or-one segment; elsewhere exactly one.
-        tailMax = 1;
-      } else {
-        fixed.push(undefined);
-      }
-    }
-  }
-  shape = { fixed, tailMin, tailMax };
-  _shapeCache.set(entry, shape);
-  return shape;
-}
-
-/**
- * Whether two canonical shapes share at least one concrete path.
- *
- * Value constraints only exist where both shapes have a *fixed* matcher at the
- * same position; tails match any value. Since any valid common length is at
- * least `max(fixedA, fixedB)`, every shared fixed position is always in range,
- * so the value check is independent of the chosen length.
- */
-function _shapesOverlap(a: RouteShape, b: RouteShape): boolean {
-  const fa = a.fixed.length;
-  const fb = b.fixed.length;
-  const common = fa < fb ? fa : fb;
-  for (let k = 0; k < common; k++) {
-    if (!_segmentsCanOverlap(a.fixed[k], b.fixed[k])) return false;
-  }
-  // Total-length ranges must intersect.
-  const lo = Math.max(fa + a.tailMin, fb + b.tailMin);
-  const hi = Math.min(fa + a.tailMax, fb + b.tailMax);
-  return lo <= hi;
-}
-
-/**
- * Whether two single-segment matchers (`string` static | `RegExp` | `undefined`
- * any) can match a common value.
- *
- * `static`/`static` and `static`/`regex` are decided precisely. Any comparison
- * involving `any`, and every `regex`/`regex` pair, is over-approximated to
- * `true` (the conservative "may overlap" default — regex intersection is
- * undecidable in general).
- */
-function _segmentsCanOverlap(
-  x: string | RegExp | undefined,
-  y: string | RegExp | undefined,
-): boolean {
-  if (typeof x === "string") {
-    return typeof y === "string" ? x === y : y instanceof RegExp ? y.test(x) : true;
-  }
-  if (x instanceof RegExp && typeof y === "string") return x.test(y);
-  return true;
-}
-
-/**
- * Whether any query shape can match the static segment `key` at `depth`
- * (either via its fixed matcher there, or via its any-value tail).
- */
-function _mayMatchAt(query: RouteShape[], depth: number, key: string): boolean {
-  for (const q of query) {
-    if (depth < q.fixed.length) {
-      const m = q.fixed[depth];
-      if (m === undefined || (typeof m === "string" ? m === key : m.test(key))) return true;
-    } else if (depth - q.fixed.length < q.tailMax) {
-      return true;
-    }
-  }
-  return false;
 }
